fix(주정윤): 본인 코드 에러 해결
- 팀원 모두 push하면 에러 생길지도
