--- src/main/java/com/back/admin/domain/model/OrderProduct.java
@@ -1,33 +1,0 @@
-package com.back.admin.domain.model;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-@NoArgsConstructor
-@Table(name = "order_product")
-public class OrderProduct {  // 주문 상품 정보
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "order_product_id")
-    private Integer id;        // 상품 코드
-
-    @ManyToOne
-    @JoinColumn(name = "product_id")
-    private Product product;
-
-    @ManyToOne
-    @JoinColumn(name = "order_id")
-    private Order order;
-
-    @Column(name = "order_quantity")
-    private Integer orderQuantity;
-
-    @Column(name = "price")
-    private Integer price;
-}
