--- src/main/java/com/back/admin/domain/model/OrderStatus.java
@@ -1,16 +1,0 @@
-package com.back.admin.domain.model;
-
-import lombok.Getter;
-
-@Getter
-public enum OrderStatus {
-    PAID("결제완료"),
-    SHIPPED("배송중"),
-    DELIVERED("배송완료");
-
-    private final String description;
-
-    OrderStatus(String description) {
-        this.description = description;
-    }
-}
