--- src/main/java/com/back/admin/domain/model/ProductDetail.java
@@ -1,23 +1,0 @@
-package com.back.admin.domain.model;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "product_detail")
-public class ProductDetail {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "detail_id")
-    private Integer id;
-
-    @OneToOne
-    @JoinColumn(name = "product_id", unique = true)
-    private Product product;
-
-    private String description;
-}
