--- src/main/java/com/back/admin/dto/OrderProductDto.java
@@ -1,8 +1,0 @@
-package com.back.admin.dto;
-
-public record OrderProductDto(
-        Integer productId,
-        String productName,
-        Integer quantity,
-        Long price
-) {}
