--- src/main/java/com/back/admin/repository/ProductRepository.java
@@ -1,9 +1,0 @@
-package com.back.admin.repository;
-
-import com.back.admin.domain.model.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-}
