+++ src/main/java/com/back/domain/order/dto/OrderItemDto.java
@@ -1,0 +1,11 @@
+package com.back.domain.order.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class OrderItemDto {
+    private int productId;
+    private int quantity;
+}
