+++ src/main/java/com/back/domain/order/dto/OrderProductDto.java
@@ -1,0 +1,8 @@
+package com.back.domain.order.dto;
+
+public record OrderProductDto(
+        Integer productId,
+        String productName,
+        Integer quantity,
+        Long price
+) {}
