+++ src/main/java/com/back/domain/order/entity/OrderStatus.java
@@ -1,0 +1,16 @@
+package com.back.domain.order.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum OrderStatus {
+    PAID("결제완료"),
+    SHIPPED("배송중"),
+    DELIVERED("배송완료");
+
+    private final String description;
+
+    OrderStatus(String description) {
+        this.description = description;
+    }
+}
