+++ src/main/java/com/back/domain/order/repository/OrdersRepository.java
@@ -1,8 +1,13 @@
 package com.back.domain.order.repository;
 
-import com.back.admin.domain.model.OrderStatus;
+import com.back.domain.order.entity.OrderStatus;
+import com.back.domain.order.entity.Orders;
 import org.springframework.data.jpa.repository.EntityGraph;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface OrdersRepository extends JpaRepository<Orders, Integer> {
