+++ src/main/java/com/back/domain/order/service/OrderService.java
@@ -1,15 +1,24 @@
 package com.back.domain.order.service;
 
-import com.back.admin.domain.model.OrderStatus;
-import com.back.order.dto.OrderForm;
-import com.back.order.dto.OrderItemDto;
+import com.back.domain.order.dto.OrderDetailDto;
+import com.back.domain.order.dto.OrderProductDto;
+import com.back.domain.order.entity.OrderStatus;
+import com.back.domain.order.entity.Orders;
+import com.back.domain.order.repository.OrdersRepository;
+import com.back.domain.product.entity.Product;
+import com.back.domain.product.repository.ProductRepository;
+import com.back.domain.order.dto.OrderForm;
+import com.back.domain.order.dto.OrderItemDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalTime;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -52,10 +61,24 @@ public class OrderService {
         return ordersRepository.findByStatus(status);
     }
 
+    public OrderDetailDto getOrderDetail(Integer orderId) {
+        Orders order = ordersRepository.findWithProductsById(orderId)
+                .orElseThrow(() -> new IllegalArgumentException("주문 없음: " + orderId));
+
+        List<OrderProductDto> products = order.getOrderProducts().stream()
+                .map(op -> new OrderProductDto(op.getProduct().getProductId(), op.getProduct().getName(),
+                        op.getQuantity(), op.getProduct().getPrice().longValue()))
+                .collect(Collectors.toList());
+
+        boolean canCancel = canCancelShipment(order);
+        return new OrderDetailDto(order.getId(), order.getCustomerName(), order.getEmail(), order.getAddress(), order.getZipcode(), order.getOrderDate(),
+                order.getStatus().name(), order.getTotalPrice(), products, canCancel);
+    }
+
     // --- 배송 ---
     public Orders shipOrder(Integer orderId) {
         Orders order = findOrderById(orderId);
-        if (order.getStatus() != OrderStatus.PAYMENT_COMPLETED) {
+        if (order.getStatus() != OrderStatus.PAID) {
             throw new IllegalStateException("배송 불가 상태");
         }
         order.setStatus(OrderStatus.SHIPPED);
@@ -63,15 +86,16 @@ public class OrderService {
         return order;
     }
 
+    // --- 배송 취소 ---
     public Orders cancelShipment(Integer orderId) {
         Orders order = findOrderById(orderId);
         if (order.getStatus() != OrderStatus.SHIPPED) {
             throw new IllegalStateException("취소 불가");
         }
         if (!canCancelShipment(order)) {
-            throw new IllegalStateException("이미 취소 불가 시간");
+            throw new IllegalStateException("취소 불가 시간");
         }
-        order.setStatus(OrderStatus.PAYMENT_COMPLETED);
+        order.setStatus(OrderStatus.PAID);
         order.setShippedAt(null);
         return order;
     }
@@ -85,6 +109,7 @@ public class OrderService {
         ordersRepository.saveAll(list);
     }
 
+    // 취소 가능 여부 판단
     private boolean canCancelShipment(Orders order) {
         if (order.getStatus() != OrderStatus.SHIPPED) return false;
         OffsetDateTime now = OffsetDateTime.now(KST_OFFSET);
@@ -99,3 +124,15 @@ public class OrderService {
         return now.isBefore(scheduledDelivery);
     }
 }
+
+/**
+ * 설명(정책 정리)
+ *
+ * 관리자가 배송하기 클릭 → shippedAt = 현재 시각 기록.
+ *
+ * canCancelShipment는 shippedAt 기준으로 해당 배송에 적용될 14:00(같은날 또는 다음날) 을 계산한다. 그 14:00 이전이면 취소 허용. (즉, 배송 처리가 실제로 시행되기 전까지만 취소 가능)
+ *
+ * deliverAllShippedOrders()는 매일 14:00에 호출되어 SHIPPED → DELIVERED 변환.
+ *
+ * 위 규칙은 "배송이 매일 14:00에 일괄 처리된다"는 전제하에 자연스럽게 작동함. (테스트 시 cron을 매분으로 바꿔서 동작 확인 가능)
+ */
