+++ src/main/java/com/back/domain/product/dto/ProductDetailDto.java
@@ -1,0 +1,16 @@
+package com.back.domain.product.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ProductDetailDto {
+    private Integer productId;
+    private String origin;
+    private String flavorAroma;
+    private String feature;
+    private Integer price;   // 팝업에 표시할 가격
+}
