+++ src/main/java/com/back/domain/product/repository/ProductRepository.java
@@ -1,5 +1,7 @@
 package com.back.domain.product.repository;
 
+import com.back.domain.product.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
