--- src/main/java/com/back/jongwoo/Product.java
@@ -1,47 +1,0 @@
-package com.back.jongwoo;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "product")
-public class Product {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "product_id")
-    private Integer productId;
-
-    @Column(nullable = false, length = 150)
-    private String name;
-
-    @Column(nullable = false)
-    private Integer price;
-
-    @Column(nullable = false)
-    private Integer quantity;
-
-    @Lob
-    private String description; // NULL 허용
-
-    @Column(name = "image_url")
-    private String imageUrl;    // NULL 허용
-
-    @Column(nullable = false)
-    private Integer stock;      // NOT NULL (덤프 기준 default 0)
-
-    // --- Getter/Setter ---
-    public Integer getProductId() { return productId; }
-    public void setProductId(Integer productId) { this.productId = productId; }
-    public String getName() { return name; }
-    public void setName(String name) { this.name = name; }
-    public Integer getPrice() { return price; }
-    public void setPrice(Integer price) { this.price = price; }
-    public Integer getQuantity() { return quantity; }
-    public void setQuantity(Integer quantity) { this.quantity = quantity; }
-    public String getDescription() { return description; }
-    public void setDescription(String description) { this.description = description; }
-    public String getImageUrl() { return imageUrl; }
-    public void setImageUrl(String imageUrl) { this.imageUrl = imageUrl; }
-    public Integer getStock() { return stock; }
-    public void setStock(Integer stock) { this.stock = stock; }
-}
