--- src/main/java/com/back/jongwoo/ProductDetail/ProductDetail.java
@@ -1,29 +1,0 @@
-package com.back.jongwoo.ProductDetail;
-
-import com.back.jongwoo.Product;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "product_detail")
-@Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
-public class ProductDetail {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "detail_id")
-    private Integer detailId;
-
-    @OneToOne
-    @JoinColumn(name = "product_id")
-    private Product product;
-
-    @Column(columnDefinition = "TEXT")
-    private String origin;
-
-    @Column(columnDefinition = "TEXT")
-    private String flavorAroma;
-
-    @Column(columnDefinition = "TEXT")
-    private String feature;
-}
