--- src/main/java/com/back/jongwoo/dto/ProductDetailDto.java
@@ -1,16 +1,0 @@
-package com.back.jongwoo.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ProductDetailDto {
-    private Integer productId;
-    private String origin;
-    private String flavorAroma;
-    private String feature;
-    private Integer price;   // 팝업에 표시할 가격
-}
