--- src/main/java/com/back/jongwoo/repository/ProductRepository.java
@@ -1,7 +1,0 @@
-package com.back.jongwoo.repository;
-
-import com.back.jongwoo.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-}
