--- src/main/java/com/back/jongwoo/service/ProductService.java
@@ -1,9 +1,0 @@
-package com.back.jongwoo.service;
-
-import com.back.jongwoo.dto.ProductDetailDto;
-
-public interface ProductService {
-    ProductDetailDto getProductDetail(Integer id);
-}
-
-// Repository → Service → Controller 흐름 생성
