--- src/main/java/com/back/order/dto/OrderItemDto.java
@@ -1,11 +1,0 @@
-package com.back.order.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class OrderItemDto {
-    private int productId;
-    private int quantity;
-}
