--- src/main/java/com/back/order/repository/OrderRepository.java
@@ -1,8 +1,0 @@
-package com.back.order.repository;
-
-import com.back.order.entity.Orders;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Orders, Integer> {
-
-}
