--- src/main/java/com/back/product/entity/ProductDetail.java
@@ -1,17 +1,0 @@
-package com.back.product.entity;
-
-import jakarta.persistence.*;
-
-@Entity
-public class ProductDetail {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id")
-    private Product product;
-
-    private String description;
-}
