--- src/main/java/com/back/product/repository/ProductRepository.java
@@ -1,8 +1,0 @@
-package com.back.product.repository;
-
-import com.back.product.entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-
-}
